Product and Inventory service and controller methods

+++ src/main/java/com/springboot/backend/gestiondeinventarios/repository/InventoryRepository.java
@@ -3,5 +3,6 @@ package com.springboot.backend.gestiondeinventarios.repository;
 import com.springboot.backend.gestiondeinventarios.models.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+
 }
